+++ src/modules/ipsec.c
@@ -36,14 +36,14 @@
  * @param co Pointer to T50 configuration structure.
  * @param size Pointer to packet size (updated by the function).
  */
-void ipsec(const struct config_options *const __restrict__ co, size_t *size)
+void ipsec(const struct config_options *const __restrict__ co, uint32_t *size)
 {
   /* IPSec AH Integrity Check Value (ICV). */
 #define IP_AH_ICV (sizeof(uint32_t) * 3)
 
-  size_t length,
-         esp_data,    /* IPSec ESP Data Encrypted (RANDOM). */
-         counter;
+  uint32_t length,
+           esp_data,    /* IPSec ESP Data Encrypted (RANDOM). */
+           counter;
 
   /* Packet. */
   memptr_t buffer;
